+++ example.c
@@ -12,40 +12,42 @@
 // ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
 // OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
-// Version: 2018-09-05
+// Version: 2018-10-08
 
 #define CONFIG_IMPLEMENTATION    // Include implementation in this file
 //#define CONFIG_VERBOSE         // Print unhandled items, missing items, and multiples to stderr
 //#define CONFIG_DEBUG           // Only for library testing
 #include "Config.h"
 
-int callbackResult = 0;
-
-// Callback checks if the given value is between 5 and 10 or exits
-void test_callback(void* sender, const char* key, const char* value)
-{ 
+// Callback checks if the given value is between 1 and 10 or exits
+void exampleCallback(void* sender, const char* key, const char* value)
+{
     printf("Inside callback for key: %s with value %s\n", key, value);
 
-    int result = atoi(value);
-    if(result < 1|| result > 10) {
-        fprintf(stderr, "Error parsing `%s`: Value `%d` is out of range [1, 10]\n", key, result);
-        exit(1);
-    }
+    // Save the value to the sender object (in this example: &callbackResult)
+    int* result = sender;
+    *result = atoi(value);
 
-    callbackResult = result;
+    // Exit program if it is out of bounds to avoid runtime errors
+    if(*result < 1|| *result > 10) {
+        fprintf(stderr, "Error parsing `%s`: Value `%d` is out of range [1, 10]\n", key, *result);
+        exit(EXIT_FAILURE);
+    }
 }
 
 int main(int argc, const char **argv)
 {
-    // Variables that will be set using Config_load().
-    // They keep their initial value, if the config is not given.
-    // String should be NULL and will be allocated inside Config.h.
+    // Declare variables that will be set using Config_load().
+    // They keep their initial value, if their ConfigItem does not match.
+    // Strings should be always be NULL and will be allocated inside Config.h.
+    // A default value for a string can be set afterwards by checking for NULL.
     bool cBool = 0;
     int cInt = 0;
     double cDouble = 0.f;
-    const char *cString = NULL;
-    const char *cKey1 = NULL;
-    const char *cKey2 = NULL;
+    const char* cString = NULL;
+    const char* cKey1 = NULL;
+    const char* cKey2 = NULL;
+    int callbackResult = 0;
 
     // Load values from configuration file to variables
     Config_load("example.txt", (ConfigItem[]) {
@@ -55,7 +57,7 @@ int main(int argc, const char **argv)
         CONFIG_INT("anInt", &cInt),
         CONFIG_STRING("namespace.key1", &cKey1),
         CONFIG_STRING("namespace.key2", &cKey2),
-        CONFIG_CALLBACK("otherspace.key1", &test_callback, NULL),
+        CONFIG_CALLBACK("otherspace.key1", &exampleCallback, &callbackResult),
         CONFIG_END()
     });